This is a wide-sweeping set of changes. Most of the changes are trivial: TIdSipLocation is now in its own unit, IdSipLocation; since this class is referenced in a great many places, many units have had their uses clauses updated.
Other than that, we now almost completely ignore the TIdSipAbstractCore.Contact property: when we send a request that might establish a dialog (or respond to that kind of message) we change the host and port of the first Contact URI in the message. First, when you configure the UA, you can set certain "mapped routes" - routes through NATs - for certain destinations (which may include the default route). When the message is sent, the UA asks the OS for the most appropriate local address to use (because your machine may have one or more LAN IPs, it might have a VPN IP, it might have an public nternet IP). Should the OS wish to use the default route to route the message, the UA will consult your set of mapped routes. If a match is found, then the address/port specified in that mapped route will be used in the Contact URI. Typically this will be the external IP of your NAT, and the port that will send SIP messages to your UA.