This commit removes every single abstract method: all these methods are now virtual and either call RaiseAbstractError or Fail, depending on whether they're part of a production class or a test class.
You might wonder why all this palaver. Well, when one uses factories to create objects (or registries to map strings to types), Delphi cannot, at compile time, determine whether a method call is a call to an abstract method. The only way we can know we've tried to call an abstract method is by receiving EAbstractErrors. That's fine. What's intolerable is that these EAbstractErrors give NO information as to WHAT method is abstract. RuntimeSafety's RaiseAbstractError tells us the name of the class that failed to override the "abstract" method, and the method's name.