TIdTimerQueues may now schedule Wait objects in other TimerQueues, if they know the other TimerQueues' registered IDs.
This allows a style of concurrency that's like asynchronous message-passing, only instead of messages we send _behaviours_. (I'm not sure if this is a sane idea, but anyway. Let's see where it takes us.) Specifically, we have threads of execution, and scheduled closures that close over objects or over registered objects. (Registered objects allow a "weak reference" model where you can give a handle of an object, and the user of the handle knows that that object may or may not still be instantiated.)