PR 854: Messages sent from non-standard setups can have wrong Via headers.
If your SIP stack doesn't use any standard ports, and you send a message on an outbound TCP connection, the message will have not have the correct topmost Via's port (one of the ports on which the stack listens).

You MUST use a ip/port/transport on which you are willing to accept messages.